Policy for users/admin

+++ app/Http/Controllers/ListingController.php
@@ -2,8 +2,13 @@
 
 namespace App\Http\Controllers;
 
+use App\Models\User;
 use App\Models\Listing;
 use Illuminate\Http\Request;
+use Illuminate\Auth\Access\Response;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
+use Symfony\Component\HttpKernel\Exception\HttpException;
 
 class ListingController extends Controller
 {
@@ -23,8 +28,13 @@ class ListingController extends Controller
     /**
      * Show the form for creating a new resource.
      */
-    public function create()
+    public function create(Request $request, Listing $listing)
     {
+        // if (Auth::user()->cannot('create', Listing::class)) {
+        //     throw new HttpException(403, 'У вас нет прав для создания нового объявления.');
+        // }
+        // Gate::authorize('update', $listing);
+
         return inertia('Listing/Create');
     }
 
@@ -33,6 +43,7 @@ class ListingController extends Controller
      */
     public function store(Request $request)
     {
+        // Gate::authorize('create', Listing::class); // Только авторизованные пользователи могут создавать листинг
         $validatedData = $request->validate([
             'beds' => 'required|integer|min:0|max:20',
             'baths' => 'required|integer|min:0|max:20',
@@ -55,6 +66,11 @@ class ListingController extends Controller
      */
     public function show(Listing $listing)
     {
+
+        // if (Auth::user()->cannot('view', $listing)) {
+        //     abort(403);
+        // }
+
         return inertia(
             'Listing/Show',
             [
@@ -68,6 +84,8 @@ class ListingController extends Controller
      */
     public function edit(Listing $listing)
     {
+        Gate::authorize('update', $listing);
+
         return inertia(
             'Listing/Edit',
             [
@@ -103,6 +121,8 @@ class ListingController extends Controller
      */
     public function destroy(Listing $listing)
     {
+        Gate::authorize('delete', $listing); // Проверка прав на удаление
+
         $listing->delete();
 
         return redirect()->back()
